Move moverow into a directory
